Refactor home page to support metadata
